procedural handler macro & generic fn input
