+++ core/guest/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "assemblylift-core-guest"
-version = "0.2.0"
+version = "0.4.0-alpha.0"
 description = "AssemblyLift core WASM guest library"
 authors = ["Akkoro and the AssemblyLift contributors <assemblylift@akkoro.io>"]
 edition = "2018"
@@ -10,3 +10,11 @@ readme = "README.md"
 
 [dependencies]
 paste = "0.1.12"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+assemblylift-core-guest-macros = { version = "0.0.0", path = "./macros" }
+
+## TESTING
+#direct-executor = "0.3.0"
+#assemblylift-core-io-guest = { version = "0.3", path = "../io/guest" }
+#serde_json = "1"
