+++ core/guest/macros/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "assemblylift-core-guest-macros"
+version = "0.0.0"
+description = "Procedural macros for assemblylift-core-guest"
+authors = ["Akkoro and the AssemblyLift contributors <assemblylift@akkoro.io>"]
+edition = "2018"
+license-file = "../../../LICENSE.md"
+repository = "https://github.com/akkoro/assemblylift"
+readme = "README.md"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+proc-macro2 = "1.0"
+quote = "1.0"
+syn = { version = "1.0", features = ["full"] }
