+++ providers/aws-lambda/host/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "assemblylift-awslambda-host"
-version = "0.3.1"
+version = "0.4.0-alpha.0"
 description = "AssemblyLift AWS Lambda runtime"
 authors = ["Akkoro and the AssemblyLift contributors <assemblylift@akkoro.io>"]
 edition = "2018"
