+++ providers/aws-lambda/host/src/runtime.rs
