+++ providers/aws-lambda/host/src/wasm.rs
